+++ docs/release-notes.rst
@@ -23,13 +23,13 @@ Release Notes
 The client and server are released together.  The client is deployed as a jar file to a
 Maven Nexus repository. The server is deployed as a Docker image to a Docker registry.
 
-Version 1.14.0, 19 February 2018
+Version 1.14.0, 21 February 2018
 -------------------------------
 
 * Add search-by-date method for federation subscription update (ACUMOS-61)
 * Add peer group, solution group and mapping features for federation access control (ACUMOS-62)
-* Add feature for User notification preference and user notification (ACUMOS-166) 
 * Refactor to drop code-name database tables (ACUMOS-144)
+* Add feature for user notification preference and user notification (ACUMOS-166) 
 * Assert unique constraint on peer subjectName attribute (ACUMOS-214)
 * Revise peer status code/name value set (ACUMOS-215)
 * Add new toolkit type code for ONAP (ACUMOS-232)